pyproject.toml
[tool.poetry]
name = "spatiotemporal-postprocessing"
# Keep only Major and Minor version in this configuration,
# the patch version is calculated when releasing the library
version = "1.0"
description = "Project on spatiotemporal modelling using GNNs to postprocess ensemble forecasts."
authors = ["Michele Cattaneo <michele.cattaneo@meteoswiss.ch>"]
readme = "README.md"
include = ["HISTORY.rst"]

[[tool.poetry.source]]
name = "pyg"
url = "https://data.pyg.org/whl/torch-2.5.0+cu121.html"
priority = "supplemental"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
torch = "2.5.1"
numpy = "^2.2.2"
torch-spatiotemporal = "^0.9.5"
torch-geometric = "^2.6.1"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
mlflow = "^2.19.0"
xarray = "2025.1.1"
torch-scatter = {version = "^2.1.2", source = "pyg"}
torch-sparse = {version = "^0.6.18", source = "pyg"}
zarr = "^3.0.1"
scoringrules = "^0.7.1"
seaborn = "^0.13.2"
numcodecs = "^0.15.0"
cartopy = "^0.24.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.0"
pydata-sphinx-theme = "^0.15.2"
pylint = "^3.0.2"
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
sphinx = "^8.0.2"
sphinx-autoapi = "^3.1.0"
yapf = "^0.40.2"
